Modified mist to accept the timeout as an optional argument:
I have modified the funcion 'mist_cmdline_options_handle'.
Now, this functions suppose that after the specification of the algorithm should appear the name of the file and that there is an optional argument 'timeout' that should appear followed by an integer which represent the timeout that the user wants to establish (in seconds).

I have also updated the function 'print_help'. I think that I did it wrong but I don't know how to specify that an argument is optional